% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-datalist.R
\docType{data}
\name{AAACF}
\alias{AAACF}
\title{Atom-Centred Fragments Descriptors for 20 Amino Acids calculated by Dragon}
\value{
AAACF data
}
\description{
Atom-Centred Fragments Descriptors for 20 Amino Acids calculated by Dragon
}
\details{
This dataset includes the atom-centred fragments descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAACF)

}
