% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-datalist.R
\docType{data}
\name{AABLOSUM100}
\alias{AABLOSUM100}
\title{BLOSUM100 Matrix for 20 Amino Acids}
\usage{
data(AABLOSUM100)
}
\value{
AABLOSUM100 data
}
\description{
BLOSUM100 Matrix for 20 Amino Acids
}
\details{
BLOSUM100 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AABLOSUM100)

}
\keyword{AABLOSUM100}
\keyword{BLOSUM}
\keyword{datasets}

