% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-datalist.R
\docType{data}
\name{AABLOSUM45}
\alias{AABLOSUM45}
\title{BLOSUM45 Matrix for 20 Amino Acids}
\value{
AABLOSUM45 data
}
\description{
BLOSUM45 Matrix for 20 Amino Acids
}
\details{
BLOSUM45 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AABLOSUM45)

}
