% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-datalist.R
\docType{data}
\name{AABLOSUM50}
\alias{AABLOSUM50}
\title{BLOSUM50 Matrix for 20 Amino Acids}
\usage{
data(AABLOSUM50)
}
\value{
AABLOSUM50 data
}
\description{
BLOSUM50 Matrix for 20 Amino Acids
}
\details{
BLOSUM50 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AABLOSUM50)

}
\keyword{AABLOSUM50}
\keyword{BLOSUM}
\keyword{datasets}
