% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-datalist.R
\docType{data}
\name{AAPAM250}
\alias{AAPAM250}
\title{PAM250 Matrix for 20 Amino Acids}
\value{
AAPAM250 data
}
\description{
PAM250 Matrix for 20 Amino Acids
}
\details{
PAM250 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AAPAM250)

}
