% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-datalist.R
\docType{data}
\name{AAPAM40}
\alias{AAPAM40}
\title{PAM40 Matrix for 20 Amino Acids}
\value{
AAPAM40 data
}
\description{
PAM40 Matrix for 20 Amino Acids
}
\details{
PAM40 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AAPAM40)

}
\keyword{AAPAM40}
\keyword{PAM}
\keyword{datasets}
