% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractDrugAutocorrelationPolarizability}
\alias{extractDrugAutocorrelationPolarizability}
\title{Calculates the Moreau-Broto Autocorrelation Descriptors using Polarizability}
\usage{
extractDrugAutocorrelationPolarizability(molecules, silent = TRUE)
}
\arguments{
\item{molecules}{Parsed molucule object.}

\item{silent}{Logical. Whether the calculating process
should be shown or not, default is \code{TRUE}.}
}
\value{
A data frame, each row represents one of the molecules,
each column represents one feature. This function returns 5 columns named
\code{ATSp1}, \code{ATSp2}, \code{ATSp3}, \code{ATSp4}, \code{ATSp5}.
}
\description{
Calculates the Moreau-Broto Autocorrelation Descriptors using Polarizability
}
\details{
Calculates the ATS autocorrelation descriptor using polarizability.
}
\examples{
\donttest{
smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugAutocorrelationPolarizability(mol)
head(dat)}
}
\author{
Nan Xiao <\url{http://r2s.name}>
}
\keyword{Autocorrelation}
\keyword{Polarizability}
\keyword{extractDrugAutocorrelationPolarizability}

