% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractDrugKRComplete}
\alias{extractDrugKRComplete}
\title{Calculate the KR (Klekota and Roth) Molecular Fingerprints (in Complete Format)}
\usage{
extractDrugKRComplete(molecules, silent = TRUE)
}
\arguments{
\item{molecules}{Parsed molucule object.}

\item{silent}{Logical. Whether the calculating process
should be shown or not, default is \code{TRUE}.}
}
\value{
An integer vector or a matrix. Each row represents one molecule,
the columns represent the fingerprints.
}
\description{
Calculate the KR (Klekota and Roth) Molecular Fingerprints (in Complete Format)
}
\details{
Calculate the 4860 bit fingerprint defined by Klekota and Roth.
}
\examples{
\donttest{
smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
mol = readMolFromSmi(smi, type = 'mol')
fp  = extractDrugKRComplete(mol)
dim(fp)}
}
\author{
Nan Xiao <\url{http://r2s.name}>
}
\seealso{
\link{extractDrugKR}
}
\keyword{extractDrugKRComplete}

