% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractDrugStandard}
\alias{extractDrugStandard}
\title{Calculate the Standard Molecular Fingerprints (in Compact Format)}
\usage{
extractDrugStandard(molecules, depth = 6, size = 1024, silent = TRUE)
}
\arguments{
\item{molecules}{Parsed molucule object.}

\item{depth}{The search depth. Default is \code{6}.}

\item{size}{The length of the fingerprint bit string. Default is \code{1024}.}

\item{silent}{Logical. Whether the calculating process
should be shown or not, default is \code{TRUE}.}
}
\value{
A list, each component represents one of the molecules, each element
in the component represents the index of which element in the fingerprint is 1.
Each component's name is the length of the fingerprints.
}
\description{
Calculate the Standard Molecular Fingerprints (in Compact Format)
}
\details{
Calculate the standard molecular fingerprints.
Considers paths of a given length.
This is hashed fingerprints, with a default length of 1024.
}
\examples{
\donttest{
smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
mol = readMolFromSmi(smi, type = 'mol')
fp  = extractDrugStandard(mol)
head(fp)}
}
\author{
Nan Xiao <\url{http://r2s.name}>
}
\seealso{
\link{extractDrugStandardComplete}
}
\keyword{extractDrugStandard}

