% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractProtPSSM}
\alias{extractProtPSSM}
\title{Compute PSSM (Position-Specific Scoring Matrix) for given protein sequence}
\usage{
extractProtPSSM(seq, start.pos = 1L, end.pos = nchar(seq),
  psiblast.path = NULL, makeblastdb.path = NULL, database.path = NULL,
  iter = 5, silent = TRUE)
}
\arguments{
\item{seq}{Character vector, as the input protein sequence.}

\item{start.pos}{Optional integer denoting the start position of the
fragment window. Default is \code{1},
i.e. the first amino acid of the given sequence.}

\item{end.pos}{Optional integer denoting the end position of the
fragment window. Default is \code{nchar(seq)},
i.e. the last amino acid of the given sequence.}

\item{psiblast.path}{Character string indicating the path of the
\code{psiblast} program.
If NCBI Blast+ was previously installed in the operation system,
the path will be automatically detected.}

\item{makeblastdb.path}{Character string indicating the path of the
\code{makeblastdb} program.
If NCBI Blast+ was previously installed in the system,
the path will be automatically detected.}

\item{database.path}{Character string indicating the path of
a reference database (a FASTA file).}

\item{iter}{Number of iterations to perform for PSI-Blast.}

\item{silent}{Logical. Whether the PSI-Blast running output
should be shown or not (May not work on some Windows versions and
PSI-Blast versions), default is \code{TRUE}.}
}
\value{
The original PSSM, a numeric matrix which has
\code{end.pos - start.pos + 1} columns and \code{20} named rows.
}
\description{
Compute PSSM (Position-Specific Scoring Matrix) for given protein sequence
}
\details{
This function calculates the PSSM (Position-Specific Scoring Matrix) derived
by PSI-Blast for given protein sequence or peptides.
For given protein sequences or peptides, PSSM represents the
log-likelihood of the substitution of the 20 types of amino acids at that
position in the sequence. Note that the output value is not normalized.
}
\note{
The function requires the \code{makeblastdb} and \code{psiblast} programs
to be properly installed in the operation system or
their paths provided.

The two command-line programs are included in the NCBI-BLAST+
software package. To install NCBI Blast+, just open the NCBI FTP site
using web browser or FTP software:
\url{ftp://anonymous@ftp.ncbi.nlm.nih.gov:21/blast/executables/blast+/LATEST/}
then download the executable version of BLAST+ according to your
operation system, and compile or install the downloaded
source code or executable program.

Ubuntu/Debian users can directly use the command
\code{sudo apt-get install ncbi-blast+} to install NCBI Blast+.
For OS X users, download \code{ncbi-blast- ... .dmg} then install.
For Windows users, download \code{ncbi-blast- ... .exe} then install.
}
\examples{
\donttest{
x = readFASTA(system.file('protseq/P00750.fasta', package = 'Rcpi'))[[1]]
dbpath = tempfile('tempdb', fileext = '.fasta')
invisible(file.copy(from = system.file('protseq/Plasminogen.fasta', package = 'Rcpi'), to = dbpath))
pssmmat = extractProtPSSM(seq = x, database.path = dbpath)
dim(pssmmat)  # 20 x 562 (P00750: length 562, 20 Amino Acids)
}
}
\author{
Nan Xiao <\url{http://r2s.name}>
}
\references{
Altschul, Stephen F., et al.
"Gapped BLAST and PSI-BLAST: a new generation of protein database search programs."
\emph{Nucleic acids research} 25.17 (1997): 3389--3402.

Ye, Xugang, Guoli Wang, and Stephen F. Altschul.
"An assessment of substitution scores for protein profile-profile comparison."
\emph{Bioinformatics} 27.24 (2011): 3356--3363.

Rangwala, Huzefa, and George Karypis.
"Profile-based direct kernels for remote homology detection and fold recognition."
\emph{Bioinformatics} 21.23 (2005): 4239--4247.
}
\seealso{
\link{extractProtPSSMFeature} \link{extractProtPSSMAcc}
}
\keyword{Alignment}
\keyword{Blast}
\keyword{PSSM}
\keyword{extract}

