% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getMolFromKEGG}
\alias{getMolFromKEGG}
\title{Retrieve Drug Molecules in MOL Format from the KEGG Database}
\usage{
getMolFromKEGG(id, parallel = 5)
}
\arguments{
\item{id}{A character vector, as the KEGG drug ID.}

\item{parallel}{An integer, the parallel parameter, indicates how many
                process the user would like to use for retrieving
                the data (using RCurl), default is \code{5}.
                For regular cases, we recommend a number less than \code{20}.}
}
\value{
A length of \code{id} character vector,
each element containing the corresponding drug molecule.
}
\description{
Retrieve Drug Molecules in MOL Format from the KEGG Database
}
\details{
This function retrieves drug molecules in MOL format from the KEGG database.
}
\examples{
\donttest{
id = 'D00496'  # Penicillamine
getMolFromKEGG(id)}
}
\author{
Nan Xiao <\url{http://r2s.name}>
}
\seealso{
See \code{\link{getSmiFromKEGG}} for retrieving drug molecules
in SMILES format from the KEGG database.
}
\keyword{KEGG}
\keyword{getDrug}
\keyword{getMolFromKEGG}

