% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105-getDrugDrugBank.R
\name{getSmiFromDrugBank}
\alias{getSmiFromDrugBank}
\title{Retrieve Drug Molecules in SMILES Format from the DrugBank Database}
\usage{
getSmiFromDrugBank(id, parallel = 5)
}
\arguments{
\item{id}{A character vector, as the DrugBank drug ID.}

\item{parallel}{An integer, the parallel parameter, indicates how many
process the user would like to use for retrieving
the data (using RCurl), default is \code{5}.
For regular cases, we recommend a number less than \code{20}.}
}
\value{
A length of \code{id} character vector, each element containing
the corresponding drug molecule.
}
\description{
Retrieve Drug Molecules in SMILES Format from the DrugBank Database
}
\details{
This function retrieves drug molecules in SMILES format from the
DrugBank database.
}
\examples{
id = 'DB00859'  # Penicillamine
\donttest{
getSmiFromDrugBank(id)}
}
\seealso{
See \code{\link{getMolFromDrugBank}} for retrieving drug molecules
in MOL format from the DrugBank database.
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
\keyword{DrugBank}
\keyword{getDrug}
\keyword{getSmiFromDrugBank}
