% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-datalist.R
\docType{data}
\name{OptAA3d}
\alias{OptAA3d}
\title{OptAA3d.sdf - 20 Amino Acids Optimized with MOE 2011.10 (Semiempirical AM1)}
\value{
OptAA3d data
}
\description{
OptAA3d.sdf - 20 Amino Acids Optimized with MOE 2011.10 (Semiempirical AM1)
}
\details{
OptAA3d.sdf - 20 Amino Acids Optimized with MOE 2011.10 (Semiempirical AM1)
}
\examples{
# This example requires the rcdk package
# library('rcdk')
# optaa3d = load.molecules(system.file('sysdata/OptAA3d.sdf', package = 'Rcpi'))
# view.molecule.2d(optaa3d[[1]])  # view the first amino acid
}
\keyword{Acid}
\keyword{Amino}
\keyword{OptAA3d}

