% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/379-extractDrugShortestPath.R
\name{extractDrugShortestPath}
\alias{extractDrugShortestPath}
\title{Calculate the Shortest Path Molecular Fingerprints (in Compact Format)}
\usage{
extractDrugShortestPath(molecules, depth = 6, size = 1024, silent = TRUE)
}
\arguments{
\item{molecules}{Parsed molucule object.}

\item{depth}{The search depth. Default is \code{6}.}

\item{size}{The length of the fingerprint bit string. Default is \code{1024}.}

\item{silent}{Logical. Whether the calculating process
should be shown or not, default is \code{TRUE}.}
}
\value{
A list, each component represents one of the molecules, each element
in the component represents the index of which element in the fingerprint is 1.
Each component's name is the length of the fingerprints.
}
\description{
Calculate the Shortest Path Molecular Fingerprints (in Compact Format)
}
\details{
Calculate the fingerprint based on the shortest paths between pairs
of atoms and takes into account ring systems, charges etc.
}
\examples{
\donttest{
smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
mol = readMolFromSmi(smi, type = 'mol')
fp  = extractDrugShortestPath(mol)
head(fp)}

}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\seealso{
\link{extractDrugShortestPathComplete}
}
\keyword{extractDrugShortestPath}

