% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/201-getProtUniProt.R
\name{getSeqFromUniProt}
\alias{getSeqFromUniProt}
\title{Retrieve Protein Sequence from the UniProt Database}
\usage{
getSeqFromUniProt(id, parallel = 5)
}
\arguments{
\item{id}{A character vector, as the protein ID.}

\item{parallel}{An integer, the parallel parameter, indicates how many
process the user would like to use for retrieving
the data (using RCurl), default is \code{5}.
For regular cases, we recommend a number less than \code{20}.}
}
\value{
A list, each component contains one of the protein
represented by amino acid sequence(s).
}
\description{
Retrieve Protein Sequence from the UniProt Database
}
\details{
This function retrieves protein represented by amino acid sequence
from the UniProt database.
}
\examples{
\donttest{
id = c('P00750', 'P00751', 'P00752')
getSeqFromUniProt(id)}

}
\author{
Nan Xiao <\url{http://nanx.me}>
}
\references{
UniProt. \url{http://www.uniprot.org/}

UniProt REST API Documentation. \url{http://www.uniprot.org/faq/28}
}
\seealso{
See \code{\link{getFASTAFromUniProt}} for retrieving protein sequence
in FASTA format from the UniProt database.
}
\keyword{UniProt}
\keyword{getProt}
\keyword{getSeqFromUniProt}

