\docType{package}
\name{Rcpi-package}
\alias{Rcpi}
\alias{Rcpi-package}
\title{Compound-Protein Interaction}
\description{
  The Rcpi package aims for ... The package is developed by
  Computational Biology and Drug Design Group, Central
  South University.
}
\details{
  \tabular{ll}{ Package: \tab Rcpi\cr Type: \tab Package\cr
  Version: \tab 1.0\cr License: \tab BSD\cr }
}
\note{
  The comprehensive user's guide could be opened with
  \code{vignette('Rcpi')}, which explains each descriptor
  included in this package and corresponding usage.

  Bug reports and feature requests should be sent to
  \url{https://github.com/road2stat/Rcpi/issues}.
}
\examples{
NULL
}
\author{
  Xiao Nan <\email{road2stat@gmail.com}> Dongsheng Cao
  \email{oriental-cds@163.com} Qingsong Xu
  \email{dasongxu@gmail.com}
}
\references{
  The associated website for this package:
  \url{http://cbdd.csu.edu.cn/Rcpi/}
}
\seealso{
  See \code{\link{read.csv}} for details.
}
\keyword{bioinforamtics}
\keyword{chemogenomics}
\keyword{chemoinformatics}
\keyword{compound}
\keyword{descriptors}
\keyword{design}
\keyword{drug}
\keyword{extraction}
\keyword{feature}
\keyword{fingerprint}
\keyword{graph}
\keyword{interaction}
\keyword{network}
\keyword{package}
\keyword{protein}
\keyword{protr}
\keyword{rcdk}
\keyword{Rcpi}
\keyword{sequence}

