\docType{package}
\name{Rcpi-package}
\alias{Rcpi}
\alias{Rcpi-package}
\title{Compound-Protein Interaction with R}
\description{
  The Rcpi package provides easily accessible functions for
  retrieving molecule data from PubChem, ChEMBL, CAS, KEGG,
  DrugBank, and protein sequences from UniProt, KEGG and
  RCSBPDB. Rcpi calculates 14 types of descriptors for
  protein sequences, 50 kinds of molecular descriptors, 9
  kinds of molecular fingerprints and computing their
  similarities. Rcpi facilitates the compound-protein
  interaction and protein-protein interaction via 3 types
  of feature combination.
}
\details{
  The package is developed by Computational Biology and
  Drug Design Group, Central South University.

  \tabular{ll}{ Package: \tab Rcpi\cr Type: \tab Package\cr
  Version: \tab 1.0\cr License: \tab BSD\cr }
}
\note{
  The comprehensive user's guide could be opened with
  \code{vignette('Rcpi')}, which explains each descriptor
  included in this package and corresponding usage.

  Bug reports and feature requests should be sent to
  \url{https://github.com/road2stat/Rcpi/issues}.
}
\examples{
NULL
}
\author{
  Xiao Nan <\email{road2stat@gmail.com}> Dongsheng Cao
  \email{oriental-cds@163.com} Qingsong Xu
  \email{dasongxu@gmail.com}
}
\references{
  The associated website for this package:
  \url{http://cbdd.csu.edu.cn/Rcpi/}
}
\seealso{
  See \code{\link{Rcpi}} for details.
}
\keyword{bioinforamtics}
\keyword{chemoinformatics}
\keyword{compound}
\keyword{descriptors}
\keyword{design}
\keyword{drug}
\keyword{extraction}
\keyword{feature}
\keyword{fingerprint}
\keyword{graph}
\keyword{interaction}
\keyword{network}
\keyword{protein}
\keyword{protr}
\keyword{rcdk}
\keyword{Rcpi}
\keyword{sequence}

