% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-Rcpi-package.R
\docType{package}
\name{Rcpi-package}
\alias{Rcpi-package}
\alias{Rcpi}
\title{Toolkit for Compound-Protein Interaction in Drug Discovery}
\description{
Rcpi offers a molecular informatics toolkit with a
comprehensive integration of bioinformatics and
cheminformatics tools for drug discovery.
}
\details{
The package vignette can be opened with \code{vignette('Rcpi')}.

\tabular{ll}{
Package: \tab Rcpi\cr
Type: \tab Package\cr
License: \tab Artistic-2.0}
}
\note{
Bug reports and feature requests should be sent to
\url{https://github.com/road2stat/Rcpi/issues}.
}
