% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcpi-package.R
\docType{package}
\name{Rcpi-package}
\alias{Rcpi}
\alias{Rcpi-package}
\title{Rcpi: Molecular Informatics Toolkit for Compound-Protein Interaction in Drug Discovery}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A molecular informatics toolkit with an integration of bioinformatics and chemoinformatics tools for drug discovery.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/Rcpi/}
  \item \url{https://github.com/nanxstats/Rcpi}
  \item Report bugs at \url{https://github.com/nanxstats/Rcpi/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me}

Authors:
\itemize{
  \item Dong-Sheng Cao
  \item Qing-Song Xu \email{qsxu@csu.edu.cn}
}

}
\keyword{internal}
