\docType{package}
\name{Rcpi-package}
\alias{Rcpi}
\alias{Rcpi-package}
\title{Toolkit for Compound-Protein Interaction in Drug Discovery}
\description{
The Rcpi package offers an R/Bioconductor package
emphasizing the comprehensive integration of bioinformatics
and chemoinformatics into a molecular informatics platform
for drug discovery.
}
\details{
The comprehensive user guide could be opened with
\code{vignette('Rcpi')}, which explains the functionalities
included in this package in detail. The Rcpi package is
developed by Computational Biology and Drug Design (CBDD)
Group, Central South University
(\url{http://cbdd.csu.edu.cn/}).

\tabular{ll}{ Package: \tab Rcpi\cr Type: \tab Package\cr
Version: \tab 1.0\cr License: \tab Artistic-2.0\cr }
}
\note{
Bug reports and feature requests should be sent to
\url{https://github.com/road2stat/Rcpi/issues}.
}
\examples{
NULL
}
\author{
Nan Xiao <\email{road2stat@gmail.com}> Dongsheng Cao
<\email{oriental-cds@163.com}> Qingsong Xu
<\email{dasongxu@gmail.com}>
}
\keyword{Rcpi}

